allow `it`, `xit`, `it.skip` and `it.only` forms.
Mocha allows calling `it()` with only a test title, which is the syntax
it uses for marking a spec as pending. `xit` delegates to `it` and does
not pass along any callback. It also supports using `it.skip()` to mark
a particular test as pending.

Another modifier for tests is `it.only()` which causes the test runner
to only run that single test for the next run.