clean up user state using a subset
Fix 430.
We need to explicitly delete user set keys on the `this` context.
The prior approach tried to rebuild the entire `this` object. That
caused problems with Ember 3.13 and up because of some changes for
tracked properties. This approach infers user defined keys and deletes those and a known list.